Agrego una confirmacion antes de eliminar un proyecto
usando window.confirm() al presionar el botón de eliminar se muestra un mensaje preguntando si el usuario está seguro de eliminar el proyecto.

+++ src/app/page.tsx
@@ -80,9 +80,13 @@ export default function Principal(){
   }
 
   const eliminarProyecto = (index:number) => {
-    const nuevaLista = proyectos.filter((_, i) => i !== index)
-    setProyectos(nuevaLista) 
-    localStorage.setItem("proyectos" , JSON.stringify(nuevaLista));
+    const confirmarEliminacion = window.confirm("Quiere Eliminar el Proyecto?");
+    if (confirmarEliminacion){
+      const nuevaLista = proyectos.filter((_, i) => i !== index)
+      setProyectos(nuevaLista) 
+      localStorage.setItem("proyectos" , JSON.stringify(nuevaLista));
+    }
+
   }
 
   return (
